Centralize Sentry capture policy and fix PGCR tooltip hydration.
Move skip/drop rules into policy.ts with a query-key registry for optional Bungie probes, slim capture.ts to thin wrappers, and replace Radix tooltips on PGCR player rows with native title attributes to stop removeChild errors.
