+++ src/lib/sentry/index.ts
@@ -1,0 +1,23 @@
+export { captureClientException, captureServerException } from "./capture"
+export { beforeSendClientEvent } from "./client"
+export {
+    buildSentryContext,
+    classifyAuthError,
+    withBungieAuthFailure,
+    type BungieAuthFailureContext,
+    type SentryCaptureContext
+} from "./context"
+export {
+    shouldDropClientEvent,
+    shouldSkipAuthServerCapture,
+    shouldSkipCapture,
+    shouldSkipMutationCapture,
+    shouldSkipReactQueryCapture,
+    shouldSkipTrpcCapture
+} from "./policy"
+export {
+    OPTIONAL_BUNGIE_QUERY_PREFIXES,
+    isOptionalBungieQueryKey,
+    sentryOptionalQueryMeta
+} from "./react-query"
+export { sentrySharedOptions, shouldDropSentryEvent } from "./shared-options"
