+++ src/services/bungie/hooks/useGlobalAlerts.ts
@@ -23,7 +23,6 @@ export const useGlobalAlerts = <T = GlobalAlert[]>(
         queryKey: ["bungie", "global-alerts", params] as const,
         queryFn: ({ queryKey }) =>
             getGlobalAlerts(bungieClient, queryKey[2]).then(res => res.Response),
-        meta: { sentryCapture: false },
         ...opts
     })
 }
